56 Update to companycontact model and new traits, and update to import ordering in api routes
